1. Removed unnecessary custom model files, as code generation doesn't overwrite the CoreData class files if they exist so there isn't a need to put class behavior in a third file (only properties get overwritten each generation)
